Small fixes. Nothing working yet - investigating.

+++ FRDM-TFC/Sources/support/PID.c
@@ -11,7 +11,7 @@
 #include "support/ARM_SysTick.h"
 #include "config.h"
 #define PID_TICKER TFC_Ticker[UPTIME_TICKER]
-#define TIME (PID_TICKER/SYSTICK_FREQUENCY)
+#define TIME ( (float)PID_TICKER / (float)SYSTICK_FREQUENCY )
 
 void UpdatePID(PID_s *PID, float ref, float actual)
 {
@@ -31,7 +31,7 @@ void UpdatePID(PID_s *PID, float ref, float actual)
 	PID->integral += PID->error * dt; //accumulate error into integral
 	PID->value += (PID->integral * PID->Ki); //sum integral
 	
-	if (PID->Kd > 0)
+	if (PID->Kd > 0.0)
 	{
 		float error_diff = (PID->error - prev_error) / dt; //calculate error derivative
 		PID->value += (error_diff * PID->Kd); //sum derivative
